Configuração do sweetalert para todos os alerts

+++ webapp/assets/js/publicacoes.js
@@ -20,7 +20,7 @@ function criarPublicacao(evento) {
     }).done(function () {
         window.location = "/home";
     }).fail(function () {
-        alert('Erro ao criar publicação!');
+        Swal.fire("Ops...", "Erro ao criar publicação!", "error");
     });
 
 }
@@ -46,7 +46,7 @@ function curtirPublicacao(evento) {
         elementoClicado.addClass('text-danger');
         elementoClicado.removeClass('curtir-publicacao');
     }).fail(function () {
-        alert('Erro ao curtir publicação!');
+        Swal.fire("Ops...", "Erro ao curtir publicação!", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     })
@@ -73,7 +73,7 @@ function descurtirPublicacao(evento) {
         elementoClicado.removeClass('text-danger');
         elementoClicado.addClass('curtir-publicacao');
     }).fail(function () {
-        alert('Erro ao descurtir publicação!');
+        Swal.fire("Ops...", "Erro ao descurtir publicação!", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     })
@@ -83,7 +83,7 @@ function atualizarPublicacao(evento) {
     $(this).prop('disabled', true);
 
     const publicacaoId = $(this).data('publicacao-id');
-    
+
     $.ajax({
         url: `/publicacoes/${publicacaoId}`,
         method: "PUT",
@@ -92,9 +92,11 @@ function atualizarPublicacao(evento) {
             conteudo: $('#conteudo').val()
         }
     }).done(function () {
-        alert("Publicação editada com sucesso!");
+        Swal.fire('Sucesso!', 'Publicação editada com sucesso!', 'success').then(function () {
+            window.location = "/home";
+        })
     }).fail(function () {
-        alert('Erro ao editar publicação!');
+        Swal.fire("Ops...", "Erro ao editar publicação!", "error");
     }).always(function () {
         $('#atualizar-publicacao').prop('disabled', false);
     });
@@ -103,20 +105,30 @@ function atualizarPublicacao(evento) {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div');
-    const publicacaoId = publicacao.data('publicacao-id');
+    Swal.fire({
+        title: 'Atenção!',
+        text: 'Você tem certeza que deseja excluir esta publicação? Essa ação é irreversível!',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+        icon: 'warning'
+    }).then(function (confirmacao) {
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop('disabled', true);
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div');
+        const publicacaoId = publicacao.data('publicacao-id');
 
-    $.ajax({
-        url: `/publicacoes/${publicacaoId}`,
-        method: "DELETE"
-    }).done(function () {
-        publicacao.fadeOut("slow", function() {
-            publicacao.remove();
+        elementoClicado.prop('disabled', true);
+
+        $.ajax({
+            url: `/publicacoes/${publicacaoId}`,
+            method: "DELETE"
+        }).done(function () {
+            publicacao.fadeOut("slow", function () {
+                publicacao.remove();
+            });
+        }).fail(function () {
+            Swal.fire("Ops...", "Erro ao excluir a publicação!", "error");
         });
-    }).fail(function () {
-        alert('Erro ao excluir a publicação!');
-    });
+    })
 }
